% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\docType{data}
\name{family}
\alias{family}
\title{Family}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
https://www.cdc.gov/nchs/nhis/nhis_2017_data_release.htm
}
\usage{
family
}
\description{
Look up family data
}
\examples{
  family
}
\keyword{datasets}
