% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medYoga.R
\docType{data}
\name{medYoga}
\alias{medYoga}
\title{Meditation/Yoga Data}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
https://www.cdc.gov/nchs/nhis/nhis_2017_data_release.htm
}
\usage{
medYoga
}
\description{
Look up meditation and Yoga data
}
\examples{
  medYoga
}
\keyword{datasets}
