% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person.R
\docType{data}
\name{nhisPerson}
\alias{nhisPerson}
\title{Person}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
https://www.cdc.gov/nchs/nhis/nhis_2017_data_release.htm
}
\usage{
nhisPerson
}
\description{
Look up person data
}
\examples{
  nhisPerson
}
\keyword{datasets}
