% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refined_sleep_set.R
\docType{data}
\name{Refined_sleep_set}
\alias{Refined_sleep_set}
\title{Sample Refined Sleep set}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
https://www.cdc.gov/nchs/nhis/nhis_2017_data_release.htm
}
\usage{
Refined_sleep_set
}
\description{
Look up sleep related columns
}
\examples{
  refinedSleepSet
}
\keyword{datasets}
