% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinedSleepSet.R
\docType{data}
\name{refinedSleepSet}
\alias{refinedSleepSet}
\title{refinedSleepSet}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
https://www.cdc.gov/nchs/nhis/nhis_2017_data_release.htm
}
\usage{
refinedSleepSet
}
\description{
Look up sleep related columns
}
\examples{
  refinedSleepSet
}
\keyword{datasets}
