% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refinedSleepSetWithSleepscore.R
\docType{data}
\name{refinedSleepSetWithSleepscore}
\alias{refinedSleepSetWithSleepscore}
\title{refinedSleepSetWithSleepscore}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
https://www.cdc.gov/nchs/nhis/nhis_2017_data_release.htm
}
\usage{
refinedSleepSetWithSleepscore
}
\description{
Look up sleep related columns with sleep score
}
\examples{
  refinedSleepSetWithSleepscore
}
\keyword{datasets}
