% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.R
\docType{data}
\name{sampleAdult}
\alias{sampleAdult}
\title{Sample Adult}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
https://www.cdc.gov/nchs/nhis/nhis_2017_data_release.htm
}
\usage{
sampleAdult
}
\description{
Look up adult data
}
\examples{
  sampleAdult
}
\keyword{datasets}
